RIP-470 Xenomorphs, now easier to restart
