+++ tox.ini
@@ -25,7 +25,7 @@ commands = pytest --durations=3 {posargs: tests}
 
 [testenv:lint-py]
 # Bottom of file has Flake8 settings
-commands = flake8 {posargs:ripley config consoler.py application.py scripts tests xenomorph.py}
+commands = flake8 {posargs:ripley config consoler.py application.py scripts tests}
 deps =
     flake8==6.1.0
     flake8-builtins>=2.1.0
